+++ src/main/java/org/ayfaar/app/controllers/AdminController.java
@@ -1,0 +1,9 @@
+package org.ayfaar.app.controllers;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@Controller
+@RequestMapping("api/admin")
+public class AdminController {
+}
