+++ src/main/webapp/index.html
@@ -20,13 +20,8 @@
     <script src="lib/angular-ui-router.min.js"></script>
     <script src="lib/ui-bootstrap/ui-bootstrap-0.11.2.min.js"></script>
 
-    <script src="js/controllers/item.js"></script>
-    <script src="js/controllers/home.js"></script>
+    <script src="js/controllers/common.js"></script>
     <script src="js/controllers/term.js"></script>
-    <script src="js/controllers/category.js"></script>
-    <script src="js/controllers/paragraph.js"></script>
-    <script src="js/controllers/item-range.js"></script>
-    <script src="js/controllers/tagger.js"></script>
     <script src="js/app.js"></script>
     <script>
         app.constant('config', {useHtml5Mode: true, apiUrl: "/api/"});
