+++ src/main/webapp/js/controllers/common.js
@@ -1,0 +1,100 @@
+function CategoryController($scope, $stateParams, $api, $state) {
+
+    $scope.name = $stateParams.name;
+    document.title = $scope.name;
+    $scope.loading = true;
+
+    $api.category.get($scope.name).then(function(category){
+        $scope.loading = false;
+        copyObjectTo(category, $scope);
+        document.title = $scope.name;
+    });
+}
+
+function HomeController($scope, $state) {
+    $scope.search = function(query) {
+        if (query) {
+            $state.goToTerm(query);
+        }
+    };
+}
+
+
+function ItemController($scope, $stateParams, $state, $api) {
+    $scope.number = $stateParams.number;
+    if (!$scope.number) {
+        $state.goToHome();
+        return
+    }
+    document.title = $scope.number;
+
+    $scope.content = "Загрузка...";
+    $api.item.get($scope.number).then(function (item) {
+        copyObjectTo(item, $scope);
+    });
+    $scope.goPrev = function() {
+        $state.go($scope.previous);
+    };
+    $scope.goNext = function() {
+        $state.go($scope.next);
+    }
+}
+
+function ItemRangeController($scope, $stateParams, $api) {
+
+    $scope.from = $stateParams.from;
+    $scope.to = $stateParams.to;
+    document.title = $scope.from+"-"+$scope.to;
+    $scope.loading = true;
+
+    $api.item.getRange($scope.from, $scope.to)
+        .then(function(items){
+            $scope.loading = false;
+            $scope.items = items;
+        });
+}
+
+
+function ParagraphController($scope, $stateParams, $api, $state) {
+
+    $scope.number = $stateParams.number;
+    document.title = "§"+$scope.number;
+    $scope.loading = true;
+
+    $api.category.get("параграф:"+$scope.number)
+        .then(function(paragrapg){
+            $scope.loading = false;
+            copyObjectTo(paragrapg, $scope);
+        });
+}
+
+
+function TaggerController($scope, $stateParams, $api) {
+    $scope.$root.hideLoop = true;
+    $scope.getTags = function(){
+        $scope.loading = true;
+        $scope.terms = [];
+        $api.term.getTermsInText($scope.text).then(function(terms){
+            $scope.terms = terms;
+            $scope.loading = false;
+        });
+    };
+}
+
+function ArticleController($scope, $stateParams, $state, $api) {
+    $scope.id = $stateParams.id;
+    if (!$scope.id) {
+        $state.goToHome();
+        return
+    }
+
+    $scope.loading = true;
+
+    $api.article.get($scope.id).then(function (a) {
+        $scope.loading = false;
+        document.title = a.name;
+        copyObjectTo(a, $scope);
+    });
+}
+
+
