--- src/main/webapp/js/controllers/home.js
@@ -1,7 +1,0 @@
-function HomeController($scope, $state) {
-    $scope.search = function(query) {
-        if (query) {
-            $state.goToTerm(query);
-        }
-    };
-}
