--- src/main/webapp/js/controllers/paragraph.js
@@ -1,14 +1,0 @@
-function ParagraphController($scope, $stateParams, $api, $state) {
-
-    $scope.number = $stateParams.number;
-    document.title = "§"+$scope.number;
-    $scope.loading = true;
-    
-    $api.category.get("параграф:"+$scope.number)
-        .then(function(paragrapg){
-            $scope.loading = false;
-            copyObjectTo(paragrapg, $scope);
-        });
-}
-
-
