--- src/main/webapp/js/controllers/tagger.js
@@ -1,13 +1,0 @@
-function TaggerController($scope, $stateParams, $api) {
-    $scope.$root.hideLoop = true;
-    $scope.getTags = function(){
-        $scope.loading = true;
-        $scope.terms = [];
-        $api.term.getTermsInText($scope.text).then(function(terms){
-            $scope.terms = terms;
-            $scope.loading = false;
-        });
-    };
-}
-
-
