+++ src/main/webapp/partials/article.html
@@ -1,0 +1,21 @@
+<div>
+    <div class="block loading" ng-show="loading">Загрузка...</div>
+    <div class="block" ng-hide="loading">
+        <h1 class="current-item-title">{{name}}</h1>
+        <div class="content" ii-bind="content"></div>
+    </div>
+</div>
+<style type="text/css">
+    #article .current-item-title {
+        font-size: 24px;
+    }
+    #article .related {
+        padding-top: 30px;
+    }
+    #article .block {
+        max-width: 600px;
+    }
+    #article .content {
+        text-align: justify;
+    }
+</style>
